+++ bpdm-gate/src/main/kotlin/org/eclipse/tractusx/bpdm/gate/service/GoldenRecordUpdateServices.kt
@@ -241,6 +241,9 @@ class GoldenRecordUpdateChunkService(
             legalEntityConfidence = legalEntityConfidence?.toUpsertData() ?: throw createMappingException(BusinessPartnerDb::legalEntityConfidence, id),
             siteConfidence = siteConfidence?.toUpsertData(),
             addressConfidence = addressConfidence?.toUpsertData() ?: throw createMappingException(BusinessPartnerDb::addressConfidence, id),
+            legalEntityUpdatedAt = legalEntityUpdatedAt,
+            siteUpdatedAt = siteUpdatedAt,
+            addressUpdatedAt = addressUpdatedAt,
             scriptVariants = scriptVariants.map { businessPartnerMappings.toScriptVariantDto(it) },
             legalEntityGoldenRecordRelations = legalEntityGoldenRecordRelations.map { it.toUpsertData() },
             addressGoldenRecordRelations = addressGoldenRecordRelations.map { it.toUpsertData() },
@@ -332,6 +335,8 @@ class GoldenRecordUpdateChunkService(
         businessPartner.legalName = header.legalName
         businessPartner.legalForm = header.legalForm
         businessPartner.shortName = header.legalShortName
+        businessPartner.legalEntityUpdatedAt = header.updatedAt
+        businessPartner.addressUpdatedAt = legalEntity.legalAddress.updatedAt
         businessPartner.legalEntityConfidence?.let { update(it,  header.confidenceCriteria) }
         businessPartner.legalEntityGoldenRecordRelations.addAll(legalEntity.header.relations.map(::toEntity))
 
@@ -349,6 +354,7 @@ class GoldenRecordUpdateChunkService(
     private fun update(businessPartner: BusinessPartnerDb, site: SiteVerboseDto){
         updateStates(businessPartner.states, site.states, BusinessPartnerType.SITE)
         businessPartner.siteName = site.name
+        businessPartner.siteUpdatedAt = site.updatedAt
         businessPartner.siteConfidence?.let { update(it,  site.confidenceCriteria) }
 
         val goldenRecordVariantByCode = site.scriptVariants.associateBy { it.scriptCode }
@@ -364,6 +370,7 @@ class GoldenRecordUpdateChunkService(
         updateIdentifiers(businessPartner.identifiers, addressProperties.identifiers.map(::toEntity), BusinessPartnerType.ADDRESS)
         updateStates(businessPartner.states, addressProperties.states, BusinessPartnerType.ADDRESS)
         businessPartner.addressName = addressProperties.name
+        businessPartner.addressUpdatedAt = addressProperties.updatedAt
         businessPartner.postalAddress.addressType = addressProperties.addressType
         businessPartner.postalAddress.physicalPostalAddress = addressProperties.physicalPostalAddress.toEntity()
         businessPartner.postalAddress.alternativePostalAddress = addressProperties.alternativePostalAddress?.toEntity()