+++ bpdm-gate/src/main/kotlin/org/eclipse/tractusx/bpdm/gate/service/SharingStateService.kt
@@ -68,7 +68,6 @@ class SharingStateService(
             byUpdatedAfter(updatedAfter)
         )
         val sharingStatePage = stateRepository.findAll(spec, pageRequest)
-
         return sharingStatePage.toPageDto {
             SharingStateDto(
                 externalId = it.externalId,