+++ CHANGELOG.md
@@ -18,6 +18,7 @@ For changes to the BPDM Helm charts please consult the [changelog](charts/bpdm/C
 
 - Script variants and script codes for business partner data [#1593](https://github.com/eclipse-tractusx/bpdm/issues/1593)
 - BPDM Gate: Golden record relations to business partner output [#1630](https://github.com/eclipse-tractusx/bpdm/issues/1630)
+- BPDM Orchestrator: updatedAt timestamps for golden record task components [#1564](https://github.com/eclipse-tractusx/bpdm/issues/1564)
 
 ### Changed