+++ CHANGELOG.md
@@ -18,6 +18,12 @@ For changes to the BPDM Helm charts please consult the [changelog](charts/bpdm/C
 
 - Script variants and script codes for business partner data [#1593](https://github.com/eclipse-tractusx/bpdm/issues/1593)
 - BPDM Gate: Golden record relations to business partner output [#1630](https://github.com/eclipse-tractusx/bpdm/issues/1630)
+- BPDM Pool: Seed default Business Partner Relation Reason Code via Flyway migration [#1679](https://github.com/eclipse-tractusx/bpdm/issues/1679)
+
+
+### Added
+
+- BPDM Gate: Migration script to enforce unique constraints to fix parallel inserts with same external id being accepted in Gate [#1546](https://github.com/eclipse-tractusx/bpdm/issues/1546)
 
 ### Changed